+++ R/plot.rspec.R
@@ -18,23 +18,28 @@
 #' @param varying a numeric vector giving values for y-axis in \code{heatplot}.
 #' @param n number of bins with which to interpolate colors and \code{varying} for the
 #' heatplot.
+#' @param labels.stack a vector of labels for the stacked spectra when using \code{type = stack}.
+#' Defaults to the numeric column ID's.   
 #' @param ... additional arguments passed to plot (or image for \code{'heatmap'}).
 #'
 #' @export
 #'
 #' @examples \dontrun{
+#' 
 #' data(teal)
 #' plot(teal, type = 'overlay')
 #' plot(teal, type = 'stack')
 #' plot(teal, type = 'heatmap')}
 #'
 #' @author Chad Eliason \email{cme16@@zips.uakron.edu}
+#' @author Thomas White \email{thomas.white026@@gmail.com}
+#' 
 #' @seealso \code{\link{spec2rgb}}, \code{\link{image}}, \code{\link{plot}}
 
 # TODO: add argument for padding region between x in stack plot
 
 plot.rspec <- function(x, select = NULL, type = c("overlay", "stack", "heatmap"),
-                       varying = NULL, n = 100, ...) {
+                       varying = NULL, n = 100, labels.stack = NULL, ...) {
   type <- match.arg(type)
 
   # make wavelength vector
@@ -202,7 +207,10 @@ plot.rspec <- function(x, select = NULL, type = c("overlay", "stack", "heatmap")
 
     yend <- tail(x2, 1)
     yloc <- ymins + yend
-    axis(side = 4, at = yloc, labels = rev(select), las = 1)
-    #  abline(h=ymins, lty=3)
+    
+    if(is.null(labels.stack))
+      labels.stack <- rev(select)
+    
+    axis(side = 4, at = yloc, labels = labels.stack, las = 1)
   }
 }